feat: refactor routes to be more modular
